refactor: split Phase 1 driver .h files into .c + .h pairs
Move function implementations from simple_watchdog.h, fan.h,
clock_source.h, bootkick.h, registers.h, and interrupts.h into
corresponding .c files. Headers now contain only declarations via
drivers.h.
